🤖 refactor: rename VS Code oRPC helpers to API
