+++ vscode/src/api/discovery.ts
